:art: move custom tenant models to correct module
